pyproject.toml
[project]
name = "kube-performance-predictor"
version = "0.1.0"
description = "Master thesis project for Software: Science and Technology"
authors = [
    {name = "NicoPolazzi",email = "nicolo.polazzi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "prometheus-api-client (>=0.6.0,<0.7.0)",
    "kubernetes (>=34.1.0,<35.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "pyyaml (>=6.0,<7.0)",
]

[tool.poetry]
packages = [{include = "kpp"}]

[dependency-groups]
dev = [
    "ruff (>=0.14.4,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pandas-stubs (>=3.0.0.260204,<4.0.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "pytest (>=8.0,<9.0)",
    "pytest-cov (>=6.0,<7.0)",
    "pytest-mock (>=3.14,<4.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# Ignore `E402` (import violations)
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=kpp --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
omit = ["kpp/config.py", "kpp/**/__init__.py", "kpp/collector/sample.py", "kpp/logging_config.py", "kpp/predictor/main.py"]
